API: Removed deprecated functions, methods and props from manager and preview API
