+++ code/ui/manager/src/container/Panel.tsx
@@ -1,7 +1,9 @@
 import type { FC } from 'react';
 import React from 'react';
 import memoize from 'memoizerific';
-import { Consumer, type Combo } from '@storybook/manager-api';
+import { Consumer } from '@storybook/manager-api';
+import type { API, Combo } from '@storybook/manager-api';
+import { Addon_TypesEnum } from '@storybook/types';
 
 import { AddonPanel } from '../components/panel/Panel';
 
@@ -11,8 +13,30 @@ const createPanelActions = memoize(1)((api) => ({
   togglePosition: () => api.togglePanelPosition(),
 }));
 
+const getPanels = (api: API) => {
+  const allPanels = api.getElements(Addon_TypesEnum.PANEL);
+  const story = api.getCurrentStoryData();
+
+  if (!allPanels || !story || story.type !== 'story') {
+    return allPanels;
+  }
+
+  const { parameters } = story;
+
+  const filteredPanels: typeof allPanels = {};
+  Object.entries(allPanels).forEach(([id, panel]) => {
+    const { paramKey }: any = panel;
+    if (paramKey && parameters && parameters[paramKey] && parameters[paramKey].disable) {
+      return;
+    }
+    filteredPanels[id] = panel;
+  });
+
+  return filteredPanels;
+};
+
 const mapper = ({ state, api }: Combo) => ({
-  panels: api.getStoryPanels(),
+  panels: getPanels(api),
   selectedPanel: api.getSelectedPanel(),
   panelPosition: state.layout.panelPosition,
   actions: createPanelActions(api),
