+++ code/ui/manager/src/runtime.ts
@@ -9,17 +9,10 @@ import { CHANNEL_CREATED } from '@storybook/core-events';
 import Provider from './provider';
 import { renderStorybookUI } from './index';
 
-const { FEATURES, CONFIG_TYPE } = global;
-
 class ReactProvider extends Provider {
-  private addons: AddonStore;
+  addons: AddonStore;
 
-  private channel: Channel;
-
-  /**
-   * @deprecated will be removed in 8.0, please use channel instead
-   */
-  private serverChannel?: Channel;
+  channel: Channel;
 
   constructor() {
     super();
@@ -33,11 +26,6 @@ class ReactProvider extends Provider {
     this.addons = addons;
     this.channel = channel;
     global.__STORYBOOK_ADDONS_CHANNEL__ = channel;
-
-    if (FEATURES?.storyStoreV7 && CONFIG_TYPE === 'DEVELOPMENT') {
-      this.serverChannel = this.channel;
-      addons.setServerChannel(this.serverChannel);
-    }
   }
 
   getElements(type: Addon_Types) {
